+++ runapp.py
@@ -1,0 +1,3 @@
+from server_app import app
+
+app.run(debug=True)
