+++ server_app/app.py
@@ -1,0 +1,6 @@
+from server_app import server_app
+
+if __name__ == '__main__':
+    server_app.run(debug=True)
+
+
